Fix kernelspec install.
While the previous implementation improved kernelspec installation, it
actually did not actually allow starting one.  This is because traitlets
assumes that class names start with an uppercase letter, so
`_MetaKernelApp` must be renamed to `MetaKernelApp`.  Fortunately it
still remains private as it is only defined in `metakernel._metakernel`.

(See https://github.com/ipython/ipython/pull/4276#issuecomment-26621822)