Update _metakernel.py
add `Error_display` method that will all an error to be thrown and also display non-text objects

In the SAS Kernel, based on metakernel, there are two streams of informaiton -- the log and the list. In parsing the log you can determine if there is an error and lag the cell as such but in a log program there could be valid output (in the form of HTML) that is helpful to the user.

This commit adds a new method to not change existing behavior.